python_tts_service/requirements.txt
# FastAPI 核心依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.5.3

# 阿里云 DashScope SDK
dashscope==1.25.0

# 音频处理
pydub==0.25.1

# 工具库
python-multipart==0.0.6
aiofiles==23.2.1

# 日志和监控
structlog==24.1.0

# 测试工具
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.26.0

# 环境配置
python-dotenv==1.0.0
